Extract shared test infrastructure with security hardening
Create test_integration_common.h with shared fixtures and helpers:
- HttpsIntegrationFixture for HTTPS/SSL integration tests
- Auth plugins (TestAuthPlugin, BlockAllFirewall)
- Embedded test certificates with secure temp file creation
- SslFactoryTestProxyGuard for testing SSL factory injection

Security improvements:
- Use O_EXCL flag to prevent TOCTOU race conditions
- Set 0600 permissions on private key temp files
- Add FD_SETSIZE bounds checking in select() calls
- Add WARNING comments on test-only credentials

Code simplification:
- Remove unused SimpleTunnelProxyFixture (~125 lines)
- Remove unused SslTunnelProxyFixture (~250 lines)
- Simplify constructor initialization
- Add named constants for ports and timing values

All 144 integration tests passing.
